+++ src/Commands/AzureAD/GetAzureADUser.cs
@@ -45,22 +45,32 @@ namespace PnP.PowerShell.Commands.Principals
         [Parameter(Mandatory = false, ParameterSetName = ParameterSet_DELTA)]
         public int? EndIndex = null;
 
+        [Parameter(Mandatory = false, ParameterSetName = ParameterSet_BYID)]
+        [Parameter(Mandatory = false, ParameterSetName = ParameterSet_LIST)]
+        public SwitchParameter IgnoreDefaultProperties;
+
         protected override void ExecuteCmdlet()
         {
             if (PnPConnection.Current.ClientId == PnPConnection.PnPManagementShellClientId)
             {
                 PnPConnection.Current.Scopes = new[] { "Directory.ReadWrite.All" };
             }
+            
+            if(ParameterSpecified(nameof(IgnoreDefaultProperties)) && !ParameterSpecified(nameof(Select)))
+            {
+                throw new ArgumentException($"When providing {nameof(IgnoreDefaultProperties)}, you must provide {nameof(Select)}", nameof(Select));
+            }
+
             if (ParameterSpecified(nameof(Identity)))
             {
                 PnP.PowerShell.Commands.Model.AzureAD.User user;
                 if (Guid.TryParse(Identity, out Guid identityGuid))
                 {
-                    user = PnP.PowerShell.Commands.Utilities.AzureAdUtility.GetUser(AccessToken, identityGuid);
+                    user = PnP.PowerShell.Commands.Utilities.AzureAdUtility.GetUser(AccessToken, identityGuid, ignoreDefaultProperties: IgnoreDefaultProperties);
                 }
                 else
                 {
-                    user = PnP.PowerShell.Commands.Utilities.AzureAdUtility.GetUser(AccessToken, WebUtility.UrlEncode(Identity), Select);
+                    user = PnP.PowerShell.Commands.Utilities.AzureAdUtility.GetUser(AccessToken, WebUtility.UrlEncode(Identity), Select, IgnoreDefaultProperties);
                 }
                 WriteObject(user);
             }
@@ -71,7 +81,7 @@ namespace PnP.PowerShell.Commands.Principals
             } 
             else
             {
-                var users = PnP.PowerShell.Commands.Utilities.AzureAdUtility.ListUsers(AccessToken, Filter, OrderBy, Select, StartIndex, EndIndex);
+                var users = PnP.PowerShell.Commands.Utilities.AzureAdUtility.ListUsers(AccessToken, Filter, OrderBy, Select, IgnoreDefaultProperties, StartIndex, EndIndex);
                 WriteObject(users, true);
             }
         }